Refactor standalone components out of app.js
