+++ client-v2/src/components/hoc/require-onboarding-complete.js
@@ -1,0 +1,23 @@
+import React, { useEffect } from 'react'
+import { connect } from 'react-redux'
+import { navigate } from 'gatsby'
+
+const requireOnboardingComplete = ComposedComponent => {
+  const Onboarded = props => {
+    useEffect(() => {
+      if (props.hasCompletedOnboarding === false) {
+        return navigate('/app/onboarding')
+      }
+    }, [props.hasCompletedOnboarding])
+
+    return <ComposedComponent {...props} />
+  }
+
+  const mapStateToProps = state => ({
+    hasCompletedOnboarding: state.onboarding.completed,
+  })
+
+  return connect(mapStateToProps)(Onboarded)
+}
+
+export default requireOnboardingComplete
