+++ client-v2/src/components/login.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import SendToReddit from './oauth/send-to-reddit'
+
+const Login = () => (
+  <>
+    <h1>Login</h1>
+    <SendToReddit />
+  </>
+)
+
+export default Login
