+++ client-v2/src/components/onboarding/email.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Link } from 'gatsby'
+
+const Email = () => (
+  <div>
+    <h2>Step 1: Email</h2>
+    <Link to='/app/onboarding/frequency'>Next</Link>
+  </div>
+)
+
+export default Email
