+++ client-v2/src/components/onboarding/frequency.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Link } from 'gatsby'
+
+const Frequency = () => (
+  <div>
+    <h2>Step 2: Frequency</h2>
+    <Link to='/app/onboarding/email'>Previous</Link>
+    <Link to='/app/onboarding/sync'>Next</Link>
+  </div>
+)
+
+export default Frequency
