+++ client-v2/src/components/onboarding/sync.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Link } from 'gatsby'
+
+const Sync = () => (
+  <div>
+    <h2>Step 3: Sync</h2>
+    <Link to='/app/onboarding/frequency'>Previous</Link>
+    <Link to='/app/onboarding/success'>Submit</Link>
+  </div>
+)
+
+export default Sync
