+++ client-v2/src/components/onboarding/welcome.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Link } from 'gatsby'
+
+const Welcome = () => (
+  <div>
+    <h2>Welcome to Unearth!</h2>
+    <Link to='/app/onboarding/email'>Let's Begin</Link>
+  </div>
+)
+
+export default Welcome
