+++ client-v2/src/components/profile.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const Profile = () => <h1>Welcome, Ceddy</h1>
+
+export default Profile
